requirements.txt
# ===== Core Framework =====
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.1.0

# ===== AWS Bedrock =====
boto3>=1.34.0
botocore>=1.34.0

# ===== DSPy =====
dspy-ai>=2.4.0

# ===== LangChain Ecosystem =====
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1

# ===== Vector Store =====
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # GPU 버전 (선택사항)

# ===== GitHub Integration =====
pygithub>=2.1.1
gitpython>=3.1.40

# ===== Utilities =====
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.26.0
numpy>=1.24.0

# ===== Logging =====
loguru>=0.7.0

# ===== Testing (Optional) =====
pytest>=7.4.0
pytest-asyncio>=0.21.0

# ===== Development Tools (Optional) =====
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0